wip(padma/resolver): partial impl of new method of definition indexing
now func definitions' type won't be `HashMap<Signature, Definitions>`
it will now require the name first, then another hashmap with signature
`HashMap<String, HashMap<Signature, Definitions>>`.

This is because a single function could have multiple overloads (same name, but different
parameter types or different return types)

Using this approach makes it much easier and faster for us to index a
function and retrieve its definitions.

But this also increases the codebase complexity due to resolving functions
returning a `Vec<>` instead of a singular data structure (e.g. the
`resolve_expression` function; it doesn't return a single BlockArgument with its
blocks, it returns possible overloads of the same expression, which is
`Vec<(Blocks, BlockArgument)>`).
The caller must find by itself which overload is appropriate for their use
to prevent unecessary resolving, we use the cache for a more efficient
lookups of functions that we might've resolved before.